........
  r184639 | file | 2009-03-27 11:18:40 -0300 (Fri, 27 Mar 2009) | 2 lines
  
  Remove a cast that is not needed.
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.6.2@184650 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ bridges/bridge_softmix.c
@@ -199,7 +199,7 @@ static int softmix_bridge_poke(struct ast_bridge *bridge, struct ast_bridge_chan
 /*! \brief Function which acts as the mixing thread */
 static int softmix_bridge_thread(struct ast_bridge *bridge)
 {
-	int timingfd = (unsigned short)(unsigned long)bridge->bridge_pvt;
+	int timingfd = (unsigned long)bridge->bridge_pvt;
 
 	ast_timer_set_rate(timingfd, (1000 / SOFTMIX_INTERVAL));
